boilerplate asciidoc parser: https://github.com/metanorma/metanorma-standoc/issues/233
