+++ lib/metanorma/standoc/cleanup.rb
@@ -7,6 +7,7 @@ require_relative "./cleanup_footnotes"
 require_relative "./cleanup_ref"
 require_relative "./cleanup_asciibib"
 require_relative "./cleanup_boilerplate"
+require_relative "./cleanup_bibdata"
 require_relative "./cleanup_section"
 require_relative "./cleanup_terms"
 require_relative "./cleanup_symbols"
